+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 [Unreleased]: https://github.com/happycollision/angular-concurrency/compare/v1.3.0...HEAD
+### Changed
+- Calling cancel on a taskInstance will be a no-op if it is not running.
 
 ## [1.3.0] - 2018-10-16
 ### Added